refactor(nexus): updated integrations tests
